requirements.txt
psycopg2-binary
python-dotenv
langdetect
nltk
fuzzywuzzy
python-Levenshtein
streamlit
langchain==0.1.20
langchain-google-genai